공지사항 - 작성, 삭제 기능

+++ HealthPlanner/src/main/java/com/health/controller/Admin_NoticeController.java
@@ -3,6 +3,7 @@ package com.health.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,10 +11,10 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import com.health.domain.MessageDTO;
 import com.health.domain.PostDTO;
 import com.health.service.Admin_NoticeService;
 
+//관리자페이지 - 공지사항
 @Controller
 public class Admin_NoticeController {
 
@@ -70,13 +71,50 @@ public class Admin_NoticeController {
 		return "Admin_NoticeDetail :: #writeAnswer";
 	}
 	
+	//삭제
 	@PostMapping(value="admin_NoticeDelete.do")
 	public String adminNoticeDelete(@RequestParam(value="postId", required=false)int postId) {
+		System.out.println("삭제 페이지");
+		if(postId == 0) {
+			return "redirect:admin_Notice.do";
+		}
 		
-		boolean isDeleted = adminNoticeService.deleteNotice(postId);
+		try {
+			boolean isDeleted = adminNoticeService.deleteNotice(postId);
+			if(isDeleted == false) {
+				System.out.println("메소드 문제 : " + isDeleted);
+			}
+		} catch (DataAccessException e) {
+			System.out.println("DB처리에서 문제 : " + e);
+		} catch (Exception e) {
+			System.out.println("시스템 처리에서 문제 : " + e);
+		}
 		
+		return "redirect:admin_Notice.do";
+	}
+	
+	//작성
+	@GetMapping(value="admin_NoticeWrite.do")
+	public String adminNoticeWrite(Model model) {
+		model.addAttribute("post",new PostDTO());
 		
-		return "redirect:Admin_NoticeDetail";
+		return "Admin_NoticeWrite";
+	}
+	
+	@PostMapping(value="admin_NoticeWriting.do")
+	public String adminNoticeWriting(final PostDTO post, Model model) {
+		System.out.println("이것은 보내지는 post인스턴스" + post);
+		try{
+			boolean isInserted = adminNoticeService.writeNotice(post);
+			if(isInserted == false) {
+				System.out.println("메소드 오류 : " + isInserted);
+			}
+		} catch (DataAccessException e) {
+			System.out.println("데이터베이스 예외 : " + e);
+		} catch(Exception e) {
+			System.out.println("시스템 예외 : " + e);
+		}
 		
+		return "redirect:admin_Notice.do";
 	}
 }
